ПР11: React доработки

+++ src/utils/Api.js
@@ -56,18 +56,18 @@ class Api {
     })
   }
 
-  addLike(id) {
-    return this._request(`${this._url}/cards/${id}/likes`, {
-      method: "PUT",
-      headers: this._headers
-    })
-  }
-
-  deleteLike(id) {
-    return this._request(`${this._url}/cards/${id}/likes`, {
-      method: "DELETE",
-      headers: this._headers
-    })
+  changeLikeCardStatus(id, isLiked) {
+    if (isLiked) {
+      return this._request(`${this._url}/cards/${id}/likes`, {
+        method: "PUT",
+        headers: this._headers
+      })
+    } else {
+      return this._request(`${this._url}/cards/${id}/likes`, {
+        method: "DELETE",
+        headers: this._headers
+      })
+    }
   }
 
   setAvatar(url) {
